Fix remaining 64-bit build warnings

+++ foo_ui_columns/fcl_fonts.cpp
@@ -31,12 +31,13 @@ class FontsDataSet : public fcl::dataset {
         {
             stream_writer_memblock mem;
             g_font_manager_data.m_common_items_entry->_export(&mem, p_abort);
-            out.write_item(identifier_global_items, mem.m_data.get_ptr(), mem.m_data.get_size());
+            out.write_item(identifier_global_items, mem.m_data.get_ptr(), gsl::narrow<uint32_t>(mem.m_data.get_size()));
         }
         {
             stream_writer_memblock mem;
             g_font_manager_data.m_common_labels_entry->_export(&mem, p_abort);
-            out.write_item(identifier_global_labels, mem.m_data.get_ptr(), mem.m_data.get_size());
+            out.write_item(
+                identifier_global_labels, mem.m_data.get_ptr(), gsl::narrow<uint32_t>(mem.m_data.get_size()));
         }
         {
             stream_writer_memblock mem;
@@ -46,9 +47,11 @@ class FontsDataSet : public fcl::dataset {
             for (t_size i = 0; i < count; i++) {
                 stream_writer_memblock mem2;
                 g_font_manager_data.m_entries[i]->_export(&mem2, p_abort);
-                out2.write_item(identifier_client_entry, mem2.m_data.get_ptr(), mem2.m_data.get_size());
+                out2.write_item(
+                    identifier_client_entry, mem2.m_data.get_ptr(), gsl::narrow<uint32_t>(mem2.m_data.get_size()));
             }
-            out.write_item(identifier_client_entries, mem.m_data.get_ptr(), mem.m_data.get_size());
+            out.write_item(
+                identifier_client_entries, mem.m_data.get_ptr(), gsl::narrow<uint32_t>(mem.m_data.get_size()));
         }
     }
     void set_data(stream_reader* p_reader, t_size stream_size, t_uint32 type, fcl::t_import_feedback& feedback,
