+++ foo_ui_columns/tab_layout.h
@@ -91,7 +91,7 @@ private:
 
     void apply();
     void initialise_presets(HWND wnd);
-    void switch_to_preset(HWND wnd, unsigned index);
+    void switch_to_preset(HWND wnd, size_t index);
 
     INT_PTR on_message(HWND wnd, UINT msg, WPARAM wp, LPARAM lp);
     void on_tree_selection_change(HTREEITEM tree_item);
@@ -100,7 +100,7 @@ private:
     bool m_initialising{};
     bool m_initialised{};
     bool m_changed{};
-    unsigned m_active_preset{};
+    size_t m_active_preset{};
     std::unordered_map<HTREEITEM, LayoutTabNode::ptr> m_node_map;
     LayoutTabNode::ptr m_node_root;
     PreferencesTabHelper m_helper{IDC_TITLE1};
